+++ packages/stunner-editors/kie-wb-common-stunner/kie-wb-common-stunner-client/kie-wb-common-stunner-widgets/src/main/java/org/kie/workbench/common/stunner/client/widgets/palette/categories/items/DefinitionPaletteItemWidgetView.java
@@ -31,9 +31,9 @@ public interface DefinitionPaletteItemWidgetView extends UberElement<DefinitionP
 
         DefaultPaletteItem getItem();
 
-        void onMouseDown(int clientX,
-                         int clientY,
-                         int x,
-                         int y);
+        void onMouseDown(double clientX,
+                         double clientY,
+                         double x,
+                         double y);
     }
 }
